requirements.txt
glfw==2.5.9
PyOpenGL==3.1.7
numpy==1.24.3
Pillow==10.0.0
opencv-python==4.11.0.86
pyside6==6.9.0
pyrr==0.10.3 
pywavefront==1.5.0
